templates/websocket_template/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
python-dotenv==1.1.0
pydantic==2.11.4
